Setting up an upcoming events calendar.
The events are based on the Google Calendar controlled by our RLBot
account. Any other Google account can write to it if we invite them.

I opted not to use Google's standard embed because it's fugly and
not sensitive to client locale settings, incredibly.

See http://sugi.github.io/jquery-gcal-flow/ to understand what's going
on here. The parameters, template, and globalization stuff is all
explained there. I set up an API key for us on the RLBot Google account.